No UnitLoader UnitBank returns prefab on request Unit is now a monobehaviour
Attach Unit to prefab with the properties so it can be retrieved as a GameComponant when the unit is instanced.  The Unit is instanced as it is added to the squad by the BattleState.  No UnitLoader Needed.  And, UnitActionController can be instanced in the AssignUnit method instead of the new SpawnUnit method.